+++ R/set_node_attr_w_fcn.R
@@ -92,7 +92,7 @@
 #' # Inspect the graph's internal
 #' # node data frame
 #' graph_3 %>% get_node_df()
-#'
+#' @family Node creation and removal
 #' @export
 set_node_attr_w_fcn <- function(graph,
                                 node_attr_fcn,